esa-oauth-server/pyproject.toml
[tool.poetry]
name = "esa-oauth-server"
version = "0.1.0"
description = ""
authors = ["kawayama <kawa130307@gmail.com>"]
readme = "README.md"
packages = [
    {include = "esa_oauth_server"},
    {include = "dbmgr", from = "../dbmgr"},
]

[tool.poetry.dependencies]
python = "^3.11"
flask = "^2.3.2"
oauthlib = "^3.2.2"
requests-oauthlib = "^1.3.1"
python-dotenv = "^1.0.0"
dbmgr = {path = "../dbmgr"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
